usb-tools: cleware: iterate over devices and find suitable one
we might want to have several cleware usb-switch attached
via a hub. We now allow users to pass the device's serial
number as the second argument to the tool so that we can
talk to the correct cleware device. If serial number isn't
passed, we will open the first found one.

Signed-off-by: Felipe Balbi <felipe.balbi@nokia.com>