usb-tools: serialc: Changed time calculation to be based on the full transfer
The code is sending and receiving the requested transfer size (typically
64000 bytes for throughput tests) in increments of
MAX_USBFS_BUFFER_SIZE, which is defined to 16 KiB. The timing
calculation was only taking into account the final transfer portion.
Thus the performance measurement was not representative of the requested
transfer size but was instead measuring the requested transfer size
modulo 16 KiB. The patch fixes this by changing the time calculation to
take into account the entire transfer.

Signed-off-by: Toni Kettula <toni.kettula@nokia.com>
Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@nokia.com>