usb-tools: serialc: Changed average calculation to cover the entire test
The existing implementation calculated the average throughput by taking
after each transfer a weighted average from the most recent transfer
throughput (10% weight) and the previous average throughput (90%
weight). This quite closely follows the most recent throughput and does
not converge to a stable value even on a long test if the momentary
throughput keeps fluctuating.

This patch changes the average calculation to take into account the
entire duration of the test.

Signed-off-by: Toni Kettula <toni.kettula@nokia.com>
Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@nokia.com>